pyproject.toml
[project]
name = "tesseland"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-ninja>=1.3.0",
    "django>=5.1.7",
    "pillow>=11.1.0",
    "scikit-image>=0.25.2",
]

[tool.uv]
dev-dependencies = [
    "ipython>=9.0.2",
    "matplotlib>=3.10.1",
]
